+++ application/teamplay-external-api/src/main/kotlin/com/teamplay/api/external/request/match/CreateMatch.kt
@@ -1,0 +1,8 @@
+package com.teamplay.api.com.teamplay.api.external.request.match
+
+import com.teamplay.domain.business.match.dto.CreateMatchDTO
+
+class CreateMatch(
+    val requesterClubId: Long,
+    val createMatchDTO: CreateMatchDTO
+)
