+++ application/teamplay-external-api/src/main/kotlin/com/teamplay/api/external/request/match/CreateMatchRequest.kt
@@ -1,0 +1,6 @@
+package com.teamplay.api.com.teamplay.api.external.request.match
+
+class CreateMatchRequest(
+    val requesterClubId: Long,
+    val contact: String
+)
