+++ application/teamplay-external-api/src/main/kotlin/com/teamplay/api/external/request/match/EnterMatchResultRequest.kt
@@ -1,0 +1,13 @@
+package com.teamplay.api.com.teamplay.api.external.request.match
+
+import com.teamplay.domain.business.match.dto.DetailResultDTO
+import com.teamplay.domain.business.match.dto.IndividualResultDTO
+
+class EnterMatchResultRequest(
+    val requesterClubId: Long,
+    val matchReview: String? = null,
+    val hostScore: Int,
+    val guestScore: Int,
+    val detailResult: MutableList<DetailResultDTO>? = mutableListOf(),
+    val individualResult: MutableList<IndividualResultDTO>? = mutableListOf()
+)
