+++ application/teamplay-external-api/src/main/kotlin/com/teamplay/api/external/response/match/MatchCreateResponse.kt
@@ -1,0 +1,7 @@
+package com.teamplay.api.com.teamplay.api.external.response.match
+
+import com.teamplay.domain.database.match.entity.Match
+
+class MatchCreateResponse(
+    val match: Match
+)
