+++ application/teamplay-external-api/src/main/kotlin/com/teamplay/api/external/response/match/MatchListResponse.kt
@@ -1,0 +1,10 @@
+package com.teamplay.api.com.teamplay.api.external.response.match
+
+import com.teamplay.domain.business.match.dto.MatchListInfoDTO
+
+class MatchListResponse (
+    val currentPage: Int,
+    val totalPage: Int,
+    val filterTitle: String,
+    val matchList: MutableList<MatchListInfoDTO>
+)
