+++ application/teamplay-external-api/src/main/kotlin/com/teamplay/api/external/response/match/MatchRequestResponse.kt
@@ -1,0 +1,7 @@
+package com.teamplay.api.com.teamplay.api.external.response.match
+
+import com.teamplay.domain.database.match.entity.MatchRequest
+
+class MatchRequestResponse(
+    val matchRequest: MatchRequest
+)
