+++ application/teamplay-external-api/src/main/kotlin/com/teamplay/api/external/response/match/MatchScheduleResponse.kt
@@ -1,0 +1,7 @@
+package com.teamplay.api.com.teamplay.api.external.response.match
+
+import com.teamplay.domain.business.match.dto.MatchScheduleListDTO
+
+class MatchScheduleResponse(
+    val matchSchedule: MutableList<MatchScheduleListDTO>
+)
