+++ application/teamplay-external-api/src/main/kotlin/com/teamplay/api/external/response/match/MatchSummaryResponse.kt
@@ -1,0 +1,7 @@
+package com.teamplay.api.com.teamplay.api.external.response.match
+
+import com.teamplay.domain.business.match.dto.MatchSummaryResultDTO
+
+class MatchSummaryResponse(
+    val matchSummaryResult: MutableList<MatchSummaryResultDTO>
+)
