+++ domain/teamplay-business-domain/src/main/kotlin/com/teamplay/domain/business/match/dto/CheckAlreadyRequestMatchDTO.kt
@@ -1,0 +1,6 @@
+package com.teamplay.domain.business.match.dto
+
+class CheckAlreadyRequestMatchDTO(
+    val matchId: Long,
+    val requesterClubId: Long
+)
