+++ domain/teamplay-business-domain/src/main/kotlin/com/teamplay/domain/business/match/dto/DetailResultDTO.kt
@@ -1,0 +1,9 @@
+package com.teamplay.domain.business.match.dto
+
+import com.teamplay.domain.database.match.entity.MatchResultType
+
+class DetailResultDTO (
+    val hostScore: Int,
+    val guestScore: Int,
+    val matchResultType: MatchResultType
+)
