+++ domain/teamplay-business-domain/src/main/kotlin/com/teamplay/domain/business/match/dto/IndividualResultDTO.kt
@@ -1,0 +1,9 @@
+package com.teamplay.domain.business.match.dto
+
+import com.teamplay.domain.database.match.entity.MatchResultType
+
+class IndividualResultDTO (
+    val score: Int,
+    val receiver: String,
+    val matchResultType: MatchResultType
+)
