+++ domain/teamplay-business-domain/src/main/kotlin/com/teamplay/domain/business/match/dto/MatchDetailResultDTO.kt
@@ -1,0 +1,7 @@
+package com.teamplay.domain.business.match.dto
+
+class MatchDetailResultDTO (
+    val hostName: String,
+    val guestName: String,
+    val matchDetailResultScore: MutableList<MatchDetailResultScoreDTO>
+)
