+++ domain/teamplay-business-domain/src/main/kotlin/com/teamplay/domain/business/match/dto/MatchDetailResultScoreDTO.kt
@@ -1,0 +1,9 @@
+package com.teamplay.domain.business.match.dto
+
+import com.teamplay.domain.database.match.entity.MatchResultType
+
+class MatchDetailResultScoreDTO (
+    val matchResultType: MatchResultType,
+    val hostScore: Int,
+    val guestScore: Int
+)
