+++ domain/teamplay-business-domain/src/main/kotlin/com/teamplay/domain/business/match/dto/MatchIndividualResultDTO.kt
@@ -1,0 +1,9 @@
+package com.teamplay.domain.business.match.dto
+
+import com.teamplay.domain.database.match.entity.MatchResultType
+
+class MatchIndividualResultDTO (
+    val matchResultType: MatchResultType,
+    val score: Int,
+    val recevier: String
+)
