+++ domain/teamplay-business-domain/src/main/kotlin/com/teamplay/domain/business/match/dto/MatchScheduleListDTO.kt
@@ -1,0 +1,9 @@
+package com.teamplay.domain.business.match.dto
+
+import com.teamplay.domain.business.match.dto.type.MatchScheduleType
+
+class MatchScheduleListDTO (
+    val scheduleTitle: String,
+    val matchScheduleInfo: MutableList<MatchScheduleInfoDTO>,
+    val matchScheduleType: MatchScheduleType
+)
