+++ domain/teamplay-business-domain/src/main/kotlin/com/teamplay/domain/business/match/dto/MatchScheduleRequestDTO.kt
@@ -1,0 +1,9 @@
+package com.teamplay.domain.business.match.dto
+
+import java.util.*
+
+class MatchScheduleRequestDTO (
+    val id: Long,
+    val startDate: Date,
+    val endDate: Date
+)
