+++ domain/teamplay-business-domain/src/main/kotlin/com/teamplay/domain/business/match/dto/MatchSummaryResultDTO.kt
@@ -1,0 +1,10 @@
+package com.teamplay.domain.business.match.dto
+
+import com.teamplay.domain.business.match.dto.type.MatchClubType
+
+class MatchSummaryResultDTO (
+    val matchClubType: MatchClubType,
+    val totalScore: Int,
+    val recentlyRecord: MutableList<Boolean>,
+    val isVictory: Boolean
+)
