--- domain/teamplay-business-domain/src/main/kotlin/com/teamplay/domain/business/match/dto/MatchValidatorDTO.kt
@@ -1,8 +1,0 @@
-package com.teamplay.domain.business.match.dto
-
-class MatchValidatorDTO {
-    class CheckAlreadyRequestMatchDTO(
-        val matchId: Long,
-        val requesterClubId: Long
-    )
-}
