+++ domain/teamplay-business-domain/src/main/kotlin/com/teamplay/domain/business/match/dto/UpdateMatchRequestDTO.kt
@@ -1,0 +1,8 @@
+package com.teamplay.domain.business.match.dto
+
+import com.teamplay.domain.database.match.entity.MatchRequestStatus
+
+class UpdateMatchRequestDTO (
+    val matchRequestId: Long,
+    val matchRequestStatus: MatchRequestStatus
+)
