+++ domain/teamplay-business-domain/src/main/kotlin/com/teamplay/domain/business/match/dto/type/MatchClubType.kt
@@ -1,0 +1,5 @@
+package com.teamplay.domain.business.match.dto.type
+
+enum class MatchClubType {
+    HOST, GUEST
+}
