--- domain/teamplay-business-domain/src/main/kotlin/com/teamplay/domain/business/match/dto/type/MatchEnum.kt
@@ -1,11 +1,0 @@
-package com.teamplay.domain.business.match.dto.type
-
-class MatchEnum {
-    enum class MatchClubType {
-        HOST, GUEST
-    }
-
-    enum class MatchScheduleType {
-        THIS_SCHEDULE, NEXT_SCHEDULE, HOST, GUEST
-    }
-}
