+++ domain/teamplay-business-domain/src/main/kotlin/com/teamplay/domain/business/match/dto/type/MatchScheduleType.kt
@@ -1,0 +1,5 @@
+package com.teamplay.domain.business.match.dto.type
+
+enum class MatchScheduleType {
+    THIS_SCHEDULE, NEXT_SCHEDULE, HOST, GUEST
+}
